moved new commands into local package
